requirements.txt
fastapi
uvicorn
python-multipart
passlib[bcrypt]
pydantic
sqlalchemy
apscheduler
python-jose[cryptography]
python-dotenv
